requirements.txt
opencv-python~=4.8.1.78
numpy~=1.26.1
pytesseract~=0.3.10
Pillow~=10.1.0
pdf2image~=1.16.3
PyPDF2~=3.0.1